read

+++ src/utils/connect.h
@@ -14,11 +14,11 @@ struct cue_connect {
     int fd;
     struct bufferevent *event;
     
-    uint32_t head_len;
+    size_t head_len;
     void *head;
     
-    uint32_t buffer_max_len;
-    uint32_t buffer_len;
+    size_t buffer_max_len;
+    size_t buffer_len;
     void *buffer;
 };
